protecting if category have item in sub category and sub category have item in child category then show a message can't delete.

+++ app/Http/Controllers/Backend/CategoryController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\DataTables\CategoryDataTable;
 use Illuminate\Http\Request;
 use App\Models\Category;
+use App\Models\SubCategory;
 use Str;
 
 class CategoryController extends Controller
@@ -91,6 +92,10 @@ class CategoryController extends Controller
     public function destroy(string $id)
     {
         $category = Category::findorFail($id); 
+        $subCategory = SubCategory::where('category_id', $category->id)->count();
+        if($subCategory > 0){
+            return response(['status' => 'error', 'message' => 'This items contain, sub items for delete this you have to delete the sub items first!']);
+        }
         $category->delete();
         return response(['status' => 'success', 'message' => 'Deleted Successfully!']); 
     }
